+++ mtui/commands/reboot.py
@@ -3,7 +3,7 @@
 from logging import getLogger
 
 from ..cli.argparse import ArgumentParser
-from ..cli.completion import complete_choices
+from ..cli.completion import complete_choices, template_completion
 from ..support.messages import NoRefhostsDefinedError
 from . import Command
 
@@ -49,5 +49,8 @@ class Reboot(Command):
     def complete(state, text, line, begidx, endidx) -> list[str]:
         """Provides tab completion for the command."""
         return complete_choices(
-            [("-t", "--target")], line, text, state["hosts"].names()
+            [("-t", "--target"), *template_completion(state)],
+            line,
+            text,
+            state["hosts"].names(),
         )